Continued work on Logue System & Curio Adjustment
Added a node which resets the logue upon exiting a logue. Added a branch to the beginning of the logue function to prevent it from trying to add the dialogueui to the viewport when it's already there, so that shouldn't throw up errors anymore.
Added a gate function to DialogueBox and MonologueBox, so you can no longer trigger them without actually overlapping the boxes themselves.