Finished Logue Function
Implemented a node which stops movement inputs upon initiating a Logue, and resets it after a Logue is finished. Logue Function is, for most intents and purposes, complete now, though future tweaks, changes, and additions might be likely, idk.
Oh, i also cleaned up the node clutter on the logue function, for what it's worth.